+++ docs/src/news.md
@@ -6,8 +6,8 @@
     - **cosm** Cosinus between the columns of a matrix
     - **cosv** Cosinus between two vectors
     - **lwmlr**: k-Nearest-neighbours locally weighted multiple linear regression (kNN-LWMLR)
-    - **lwmlr_s**: kNN-LWMLR after preliminary dimension reduction
-    - **pmod** Short-cut for function 'parentmodule'
+    - **lwmlr_s**: kNN-LWMLR after preliminary (linear or non-linear) dimension reduction
+    - **pmod** Short-cut for function 'Base.parentmodule'
     - **tabdupl** Tabulate duplicated values in a vector
 
 - Modifications